refactor(#6) - 쿠폰 리포지토리 Stub 네이밍 변경, DiscountType 할인값을 BigDecimal 타입으로 변경

+++ src/main/java/com/flab/infrun/coupon/application/command/CreateCouponCommand.java
@@ -3,6 +3,7 @@ package com.flab.infrun.coupon.application.command;
 import com.flab.infrun.coupon.domain.Coupon;
 import com.flab.infrun.coupon.domain.DiscountInfo;
 import com.flab.infrun.coupon.domain.DiscountType;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record CreateCouponCommand(
@@ -17,7 +18,7 @@ public record CreateCouponCommand(
             code,
             DiscountInfo.of(
                 DiscountType.valueOf(discountType),
-                discountAmount
+                BigDecimal.valueOf(discountAmount)
             ),
             expirationAt
         );
